+++ test/support/test_migration.ex
@@ -5,6 +5,6 @@ defmodule Jido.Ecto.TestRepo.Migrations.CreateStorageTables do
 
   def change do
     require Jido.Ecto.Migrations
-    Jido.Ecto.Migrations.create_storage_tables()
+    Jido.Ecto.Migrations.create_storage_tables(version: 1)
   end
 end