+++ test/support/test_repo.ex
@@ -1,7 +1,9 @@
 defmodule Jido.Ecto.TestRepo do
   @moduledoc false
 
+  @adapter Application.compile_env(:jido_ecto, [__MODULE__, :adapter], Ecto.Adapters.SQLite3)
+
   use Ecto.Repo,
     otp_app: :jido_ecto,
-    adapter: Ecto.Adapters.SQLite3
+    adapter: @adapter
 end